# server/src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Connection Manager API"
  version: "1.6.0"
  title: "Connection Manager API"
  contact:
    email: "ramiro@modusbox.com"
host: "localhost:3001"
basePath: "/api"
...
schemes:
- "http"
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: "password"
     # The flow used by the OAuth2 security scheme. Valid values are "implicit", "password", "application" or "accessCode".
    tokenUrl: https://WSO2_IM_SERVERl:9443/oauth2/token
    scopes:
      pta: Portal Technical Administrator
      mta: Mojabox Technical Administrator
      everyone: All logged in users
security:
  - OAuth2: [ pta ] # most restrictive role
paths:
...
  /environments/{envId}/dfsps/endpoints/unprocessed:
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Returns the unprocessed endpoint items"
      operationId: "getUnprocessedEndpointItems"
      produces:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "ID of environment"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"  
...
  /environments/{envId}/dfsps/{dfspId}/endpoints:
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Returns all DFSP endpoints"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getDFSPEndpoints"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "Environment id"
        required: true
        type: "string"
      - name: "dfspId"
        in: "path"
        description: "DFSP id"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
  /environments/{envId}/dfsps/{dfspId}/endpoints/unprocessed:
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Returns the unprocessed dfsp items"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getUnprocessedDfspItems"
      produces:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "ID of environment"
        required: true
        type: "string"
      - name: "dfspId"
        in: "path"
        description: "DFSP id"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
  /environments/{envId}/dfsps/{dfspId}/endpoints/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getDFSPEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    put:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Update an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "updateDFSPEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/EndPointAny"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    delete:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "deleteDFSPEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "DfspNetworkConfig"      
  /environments/{envId}/dfsps/{dfspId}/endpoints/{epId}/confirmation:
    post:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Updates the endpoint as confirmed"
      security:
        - OAuth2: [ pta ]
      operationId: "confirmEndpointItem"
      produces:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "ID of environment"
        required: true
        type: "string"
      - name: "dfspId"
        in: "path"
        description: "DFSP id"
        required: true
        type: "string"
      - name: "epId"
        in: "path"
        description: "ID of the endpoint item"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"
  /environments/{envId}/dfsps/{dfspId}/endpoints/ingress/ips:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    post:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Adds a new IP entry to the DFSP Ingress endpoint"
      security:
        - OAuth2: [ mta, pta ]
      description: "Adds the entry with state NEW"
      operationId: "createDFSPIngressIp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "DFSP ingress IP entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "DfspNetworkConfig"      
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get the DFSP Ingress IPs"
      security:
        - OAuth2: [ mta, pta ]
      description: ""
      operationId: "getDFSPIngressIps"
      produces:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "ID of environment"
        required: true
        type: "string"
      - name: "dfspId"
        in: "path"
        description: "DFSP id"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "DfspNetworkConfig"   
  /environments/{envId}/dfsps/{dfspId}/endpoints/ingress/ips/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getDFSPIngressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    put:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Update an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "updateDFSPIngressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    delete:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "deleteDFSPIngressIpEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "DfspNetworkConfig"      
  /environments/{envId}/dfsps/{dfspId}/endpoints/ingress/urls:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    post:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Adds a DFSP Ingress URL"
      security:
        - OAuth2: [ mta, pta ]
      description: "Adds a DFSP ingress URL with state NEW"
      operationId: "createDFSPIngressUrl"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "DFSP ingress URL"
        required: true
        schema:
          $ref: "#/definitions/InputURLValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointURL"
      x-swagger-router-controller: "DfspNetworkConfig"   
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get the DFSP Ingress URLs"
      security:
        - OAuth2: [ mta, pta ]
      description: ""
      operationId: "getDFSPIngressUrls"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointURL"
      x-swagger-router-controller: "DfspNetworkConfig"   
  /environments/{envId}/dfsps/{dfspId}/endpoints/ingress/urls/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getDFSPIngressUrlEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    put:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Update an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "updateDFSPIngressUrlEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/InputURLValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    delete:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "deleteDFSPIngressUrlEndpoint"
      consumes:
      - "application/json"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "DfspNetworkConfig"
  /environments/{envId}/dfsps/{dfspId}/endpoints/egress/ips:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    post:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Adds a new IP entry to the DFSP Egress endpoint"
      security:
        - OAuth2: [ mta, pta ]
      description: "Adds the entry with state NEW"
      operationId: "createDFSPEgressIp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "DFSP egress IP entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "DfspNetworkConfig"      
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get the DFSP Egress IPs"
      security:
        - OAuth2: [ mta, pta ]
      description: ""
      operationId: "getDFSPEgressIps"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EndPointIp"
      x-swagger-router-controller: "DfspNetworkConfig"   
  /environments/{envId}/dfsps/{dfspId}/endpoints/egress/ips/{epId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "Environment id"
      required: true
      type: "string"
    - name: "dfspId"
      in: "path"
      description: "DFSP id"
      required: true
      type: "string"
    - name: "epId"
      in: "path"
      description: "Endpoint entry id"
      required: true
      type: "string"
    get:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Get an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getDFSPEgressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    put:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Update an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "updateDFSPEgressIpEndpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "endpoint entry"
        required: true
        schema:
          $ref: "#/definitions/InputIPValue"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/EndPointAny"
      x-swagger-router-controller: "DfspNetworkConfig"      
    delete:
      tags:
      - "DFSP-Endpoint-configuration"
      summary: "Delete an endpoint entry"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "deleteDFSPEgressIpEndpoint"
      consumes:
      - "application/json"
      parameters:
      - name: "envId"
        in: "path"
        description: "Environment id"
        required: true
        type: "string"
      - name: "dfspId"
        in: "path"
        description: "DFSP id"
        required: true
        type: "string"
      - name: "epId"
        in: "path"
        description: "Endpoint entry id"
        required: true
        type: "string"
      responses:
        204:
          description: "successful operation"
      x-swagger-router-controller: "DfspNetworkConfig"      
...
definitions:
...
  URLEntry:
    properties:
      url:
        type: "string"
        description: "URL"
  IPEntry:
    properties:
      address:
        type: "string"
        description: "A single IP address or a set of address using CIDR notation \
        \ IP:\
        \ a.b.c.d\
        \ where 0 <= [a,b,c,d] <= 255\
        \ CIDR:\
        \ a.b.c.d/e\
        \ where 0 <= [a,b,c,d] <= 255, and 0 <= e <= 32"
        pattern: '^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(\/([0-9]|[1-2][0-9]|3[0-2]))?$'
        example:
          "163.10.24.28/30"
      ports:
        type: "array"
        description: "List of ports or port ranges"
        items:
          type: "string"
          description: "Individual port or port range"
          example: "8000-8080"
        example:
          - "80"
          - "8000-8080"
  BaseEndpoint:
    description: Common properties to both IP and URL endpoints
    properties:
      id:
        type: "integer"
        format: "int32"
      type:
        type: "string"
        enum:
        - "IP"
        - "URL"
      direction:
        type: "string"
        enum:
        - "INGRESS"
        - "EGRESS"
      state:
        type: "string"
        enum:
        - "NEW"
        - "CONFIRMED"
        - "INVALID"
      createdDate:
        type: "string"
        format: "date-time"
      createdBy:
        type: "string"
      dfspId:
        type: "string"
  EndPointIp:
    allOf:
      - $ref: "#/definitions/BaseEndpoint"
      - type: object
        properties:
          value:
            $ref: "#/definitions/IPEntry"
  EndPointURL:
    allOf:
      - $ref: "#/definitions/BaseEndpoint"
      - type: object
        properties:
          value:
            $ref: "#/definitions/URLEntry"
  EndPointAny:
    allOf:
      - $ref: "#/definitions/BaseEndpoint"
      - type: object
        properties:
          value:
            description: It's *either* {url} OR {address and ports}. 
            allOf:
              - $ref: "#/definitions/IPEntry"
              - $ref: "#/definitions/URLEntry"
  InputIPValue:
    type: object
    properties:
      value:
        $ref: "#/definitions/IPEntry"
  InputURLValue:
    type: object
    properties:
      value:
        $ref: "#/definitions/URLEntry"
...